Sketching out a thing to manage active instrument voices.
I expect that there are a lot of problems with this.

The basic idea is the midi status messages can activate new voices,
and they can tell an existing voice to release.  Some other mechanism
would actually mark the voice as expired, but maybe having "released"
and "expired" be optionally the same thing is useful, at least in the
near term.

The POLYPHONY constant is the maximum number of voices that can be
active concurrently.  It may or may not end up being useful to have,
but the idea is if there is an instance of the audio synthesis
machinery for each potentially active voice, then this is a control to
reduce how much space is needed on the fpga.

If it ends up that pipelining can allow all 127 notes to run smoothly
and concurrently with just one or a small number of instances of the
audio synthesis machinery, then I'll likely do away with the concept.

There are two time stamps on the "voice" record, "active" and
"expired".  The idea was that these would be count from zero how long
since the respective events, but in retrospect, having a global clock
that these things just compare against instead seems more useful.

I think the next iteration will just manage an array of 127 voices and
the scheduler stuff would be some other component that hooks into
this.  I'm not really sure if that is actually a useful design,
however.